+++ src/Control/Lens/IndexedSetter.hs
@@ -26,7 +26,6 @@ module Control.Lens.IndexedSetter
 
 import Control.Lens.Indexed
 import Control.Lens.Internal
-import Control.Lens.Setter
 import Control.Lens.Type
 import Control.Monad.State.Class as State